Welcome to Academic Journal of Medical Sciences   The Academic Journal of Medical Sciences (AJMS) is an international peer-reviewed publication, published quarterly . The aim of AJMS is to provide a scientific medium of communication for researchers throughout the globe. The Journal welcomes original clinical articles as well as basic science research experiences mainly on prevalent diseases worldwide and the analysis of various worldwide health problems. Manuscripts are publishable in the form of original articles, review articles, case reports, letters to the Editor, etc.  Manuscripts should be sent through the online submission system, or by email to editor-in-chief , kindly visit the submission page and follow the Author’s Guideline before submission.
Welcome to Academic Journal of Medical Sciences

The Academic Journal of Medical Sciences (AJMS) is an international peer-reviewed publication, published quarterly . The aim of AJMS is to provide a scientific medium of communication for researchers throughout the globe. The Journal welcomes original clinical articles as well as basic science research experiences mainly on prevalent diseases worldwide and the analysis of various worldwide health problems. Manuscripts are publishable in the form of original articles, review articles, case reports, letters to the Editor, etc.  Manuscripts should be sent through the online submission system, or by email to editor-in-chief , kindly visit the submission page and follow the Author’s Guideline before submission.